% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis}
\alias{anfis}
\title{adaptive neuro fuzzy inference system}
\usage{
anfis(asset, spec, ...)
}
\arguments{
\item{asset}{a data asset}

\item{spec}{a model specification}

\item{\dots}{arguments passed to and from methods}
}
\value{
null
}
\description{
evaluate the subjective parameters of a data asset.
}
\references{
Jang, J.R. (1993). ANFIS: adaptive-network-based- fuzzy inference
system. \emph{IEEE transactions on systems, man, and cybernetics, 23}(3), 665-685
}
\seealso{
\code{\link[=dat]{dat}} for user-facing function.
}
