% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_capitals_polys.R
\name{at_capitals_polys}
\alias{at_capitals_polys}
\title{Polygons for Austria's 9 State Capitals}
\format{
An \code{sf} object with 9 rows and X columns (see below).
}
\source{
Statistik Austria: \url{https://data.statistik.gv.at/data/OGDEXT_GEM_1_STATISTIK_AUSTRIA_20250101.zip}
}
\description{
Spatial polygon data for Austria’s 9 Landeshauptstädte (state capitals),
filtered from the official Statistik Austria municipal boundaries shapefile.
}
\details{
The data was filtered using the official municipality codes (GKZ) for the following capitals:
Wien, St. Pölten, Eisenstadt, Linz, Graz, Klagenfurt am Wörthersee, Salzburg, Innsbruck, Bregenz.
}
\examples{
plot(at_capitals_polys["geometry"])

}
\seealso{
The full shapefile was obtained from Statistik Austria's open data portal.
}
\keyword{Austria}
\keyword{datasets}
\keyword{sf}
\keyword{spatial}
