% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdse_download_ndvi.R
\name{cdse_list_ndvi_products}
\alias{cdse_list_ndvi_products}
\title{List Available Copernicus NDVI Products}
\usage{
cdse_list_ndvi_products()
}
\value{
A data frame (tibble) containing product details.
}
\description{
Displays a table of available NDVI products from the Copernicus Data Space Ecosystem,
including resolution, version, and available formats.
}
