% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdse_set_credentials.R
\name{cdse_set_credentials}
\alias{cdse_set_credentials}
\title{Set Copernicus Data Space Ecosystem Credentials}
\usage{
cdse_set_credentials(profile = "copernicus", file = "~/.aws/credentials")
}
\arguments{
\item{profile}{Character. The name of the profile to read from the credentials file.
Defaults to "copernicus".}

\item{file}{Character. Path to the AWS credentials file.
Defaults to "~/.aws/credentials".}
}
\value{
A named list containing \code{access_key} and \code{secret_key}, or NULL if not found.
}
\description{
Reads AWS credentials from the standard AWS credentials file (\verb{~/.aws/credentials})
for a specific profile (defaulting to "copernicus").
}
