% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{classify}
\alias{classify}
\title{classify a given data asset}
\usage{
classify(asset)
}
\arguments{
\item{asset}{a data asset}
}
\value{
\strong{chr}: a platform and domain independent classification.
}
\description{
classify a given data asset
}
