% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribute.R
\name{contribute}
\alias{contribute}
\title{Contribute expert opinion}
\usage{
contribute(orcid, field = NULL)
}
\arguments{
\item{orcid}{ORCID id}

\item{field}{field of research to submit default data value weights for}
}
\value{
NULL - function called for its side effects
}
\description{
Contribute to open-source effort to collect domain-specific data asset
valuation quantifiers.
}
\details{
To see a list of domains see \code{dav::domains}
}
\examples{

\dontrun{
contribute("0000-0002-4343-5851")
}
}
