% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dat}
\alias{dat}
\alias{dat.default}
\alias{dat.formula}
\title{Quantify the semantic value of data assets}
\usage{
dat(asset, spec, ...)

\method{dat}{default}(asset, spec = ~test, ...)

\method{dat}{formula}(spec = ~test, asset, ...)
}
\arguments{
\item{asset}{\code{lst}|\code{chr} Data asset(s) to value. Can be an \R object
or file path.}

\item{spec}{\code{fml} a one-sided formula to specify data attributes.}

\item{\dots}{arguments to pass to or from other methods.}
}
\value{
a structure \code{dat} that contains:
\item{value}{A numeric value.}
\item{method}{A list containing the method of computation used.}
\item{context}{A list containing a description of the asset, user and
relevant software versions.}
}
\description{
A data asset is a \dfn{managed set of relata of contextual semantic value}.
Its numeric approximation, dubbed \code{dat}, uses expert opinions and
indicators.
}
\details{
\subsection{Defining data assets valuation}{

A valuation of a data asset is an approximation of its perceived value
under reasonable constraints. It is by no means an exact science.
Definitions such as \emph{data}, \emph{information}, and \emph{file} carry different
meanings in different contexts. Reasonable constraints such as
assumptions and precision have to be placed.

A data \code{asset} can refer to file locations or objects in memory.
\itemize{
\item List-like structures, specifically non-atomic classes where
\code{is.recursive} is \code{TRUE}. This includes data.frames, environments
and structures built on them.
\item File path: A path to directory or file where the user has at least
read access.
}
}

\subsection{Using data asset valuation}{

For a detailed explanation of the assumptions, limitations of the
package, refer to the \code{vignette("dva", "dva")}.
}

These are value-driving components in data asset valuation processes.

\tabular{cc}{
\strong{Concept} \tab \strong{Description}                                \cr
Accessibility\tab  placeholder description of the attribute              \cr
Accuracy     \tab  placeholder description of the attribute              \cr
Clarity      \tab  placeholder description of the attribute              \cr
Completeness \tab  placeholder description of the attribute              \cr
Conciseness  \tab  placeholder description of the attribute              \cr
Consistency  \tab  placeholder description of the attribute              \cr
Documentation\tab  placeholder description of the attribute              \cr
Encryption   \tab  placeholder description of the attribute              \cr
Frequency    \tab  placeholder description of the attribute              \cr
Integrity    \tab  placeholder description of the attribute              \cr
Longevity    \tab  placeholder description of the attribute              \cr
Objectivity  \tab  placeholder description of the attribute              \cr
Ownership    \tab  placeholder description of the attribute              \cr
Precision    \tab  placeholder description of the attribute              \cr
Relevance    \tab  placeholder description of the attribute              \cr
Scalability  \tab  placeholder description of the attribute              \cr
Scarcity     \tab  placeholder description of the attribute              \cr
Security     \tab  placeholder description of the attribute              \cr
Sensitivity  \tab  placeholder description of the attribute              \cr
Timeliness   \tab  placeholder description of the attribute              \cr
Usability    \tab  placeholder description of the attribute              \cr
Validity     \tab  placeholder description of the attribute              \cr
Volume       \tab  placeholder description of the attribute              \cr
}
}
\examples{
\dontrun{
   dat(iris)
}
}
\seealso{
\code{\link[=classify]{classify}} and \code{\link[=is.asset]{is.asset}}
give information about asset identification and categorization.
\code{\link[=sleuth]{sleuth}} makes the computation process reproducible.
}
\author{
Donald Seinen
}
