% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dav_compare.R
\name{dav_compare}
\alias{dav_compare}
\title{Compare multiple vectors}
\usage{
dav_compare(...)
}
\arguments{
\item{...}{A variable number of vectors to compare.}
}
\value{
A data frame with a column 'element' containing all unique elements
from the input vectors. For each input vector, a logical column 'in_vector_i'
is added, indicating if the element is present in that vector. An additional
column 'in_all' is TRUE if the element is present in all vectors.
}
\description{
This function takes multiple vectors as input and returns a data frame
showing the presence or absence of each unique element across all vectors.
}
\examples{
vec1 <- c("a", "b", "c")
vec2 <- c("b", "c", "d")
vec3 <- c("c", "d", "e")
dav_compare(vec1, vec2, vec3)
}
