% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{describe}
\alias{describe}
\title{get a file description based on the file}
\usage{
describe(file)
}
\arguments{
\item{file}{filepath}
}
\value{
list of file metadata descriptors
}
\description{
get a file description based on the file
}
