% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dva.R
\name{dva}
\alias{dva}
\title{assess the value of data asset}
\usage{
\special{dav::dva(asset, spec, \dots)}

\special{dav::dva(spec, asset, \dots)}
}
\arguments{
\item{asset}{\strong{list|chr}: the data asset to value. Can be a file path
or an object.}

\item{spec}{\strong{fml}: a valuation specification.}

\item{\dots}{arguments to pass to or from other methods.}
}
\value{
list of length 3: a summary statistic, the method used and context of
evaluation.
}
\description{
\code{dva} quantifies the semantic value of data assets.
}
\details{
A \emph{data asset} is a managed set of relata that carry contextual semantic
value.

The variable \var{asset} can refer to file locations or objects in memory.
\describe{
\item{\code{\var{asset}}:}{
A collection of assets. \special{dva} supports list-like structures,
specifically non-atomic classes where \code{is.recursive} is \code{TRUE}.
}
\item{\code{\var{filepath}}:}{
A path to directory or file
}
}
See the vignette at \code{vigenette('dva', 'dva')} for a detailed explanation
of the assumptions, limitations of the package.
}
\section{Data asset valuation semantics}{

A valuation of a data asset is an approximation of its perceived value. It is
by no means an exact science. Definitions of concepts such as \emph{data},
\emph{file}, \emph{information} are contested.
}

\section{Data value attributes}{

Non-divisible value drivers of data assets.

\describe{
\item{\code{\var{Completeness}}:}{
To what degree the asset is whole
}
\item{\code{\var{Relevance}}:}{
placeholder description
}
\item{\code{\var{Integrity}}:}{
placeholder description
}
\item{\code{\var{Timeliness}}:}{
placeholder description
}
\item{\code{\var{Accessibility}}:}{
easily and quickly retrievable
}
\item{\code{\var{Consistency}}:}{
placeholder description
}
\item{\code{\var{Ownership}}:}{
placeholder description
}
\item{\code{\var{Security}}:}{
placeholder description
}
\item{\code{\var{Sensitivity}}:}{
placeholder description
}
\item{\code{\var{Usability}}:}{
placeholder description
}
\item{\code{\var{Volume}}:}{
placeholder description
}
\item{\code{\var{Conciseness}}:}{
Compactness of data. Examples are compressed file or versus those that
have redundant or repetitive non-informative content.
}
\item{\code{\var{Documentation}}:}{
Metadata on the management of the asset at hand.
}
\item{\code{\var{Encryption}}:}{
A path to directory or file
}
\item{\code{\var{Longevity}}:}{
A path to directory or file
}
\item{\code{\var{Scalability}}:}{
A path to directory or file
}
\item{\code{\var{Scarcity}}:}{
A path to directory or file
}
\item{\code{\var{Frequency}}:}{
A path to directory or file
}

}
}

\seealso{
\code{\link[=is.asset]{dav::is.asset}} \code{\link[=classify]{dav::classify}}
give information about asset specification. \code{\link[=sleuth]{dav::sleuth}}
aid in making the process reproducible across \R sessions.
}
