% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_auth.R
\name{dw_auth}
\alias{dw_auth}
\alias{dw_dst}
\alias{dw_ndr}
\title{Authenticate with Datawrapper for a Specific Employer}
\usage{
dw_dst(api_key = NULL)

dw_ndr(api_key = NULL)
}
\arguments{
\item{api_key}{Optional. A character string containing the Datawrapper API key.
If provided, the key will be saved to your \code{.Renviron} file for future use.}
}
\value{
Invisibly returns the API key that was set. It also prints a
confirmation message to the console.
}
\description{
These functions manage and set the \code{DW_KEY} environment variable for the
current R session, allowing you to easily switch between different
Datawrapper accounts (e.g., for different employers).
}
\details{
The first time you use one of these functions for a specific employer, you must
provide the \code{api_key}. This key will be stored securely in your local
\code{.Renviron} file as \code{DW_KEY_DST} or \code{DW_KEY_NDR}. On subsequent uses, you
can call the function without an \code{api_key} to load the stored key for your
current R session.
}
\examples{
\dontrun{
# First time usage (stores the key):
dw_dst(api_key = "your_api_key_for_dst")

# Subsequent usage in a new R session:
dw_dst()
}
\dontrun{
# First time usage (stores the key):
dw_ndr(api_key = "your_api_key_for_ndr")

# Subsequent usage in a new R session:
dw_ndr()
}
}
\author{
Benedict Witzenberger, Robin Kohrs
}
