% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_auth.R
\name{dw_auth_key}
\alias{dw_auth_key}
\title{Helper Function to Set and Load Datawrapper API Keys}
\usage{
dw_auth_key(key_name, api_key = NULL)
}
\arguments{
\item{key_name}{The name of the environment variable to use for storage
(e.g., "DW_KEY_DST").}

\item{api_key}{The API key string, or \code{NULL}.}
}
\description{
This internal function handles the logic for storing, retrieving, and setting
employer-specific API keys.
}
\keyword{internal}
