% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_timestamp.R
\name{dw_timestamp}
\alias{dw_timestamp}
\title{Create a Formatted Timestamp String}
\usage{
dw_timestamp(x = Sys.time(), format = "de_text", tz = "")
}
\arguments{
\item{x}{A \code{POSIXct}, \code{POSIXt}, or \code{Date} object. Defaults to \code{Sys.time()}.}

\item{format}{A character string specifying the format.
Options include:
\itemize{
\item \code{"de_text"}: "9. Dezember um 7:20 Uhr" (Default)
\item \code{"de_date"}: "9. Dezember 2025"
\item \code{"de_short"}: "09.12.2025"
\item Any valid standard format string (e.g., \code{"\%Y-\%m-\%d"}).
}}

\item{tz}{Timezone specification to be used for the conversion,
if available. Defaults to the timezone attribute of \code{x} or system default.
You can pass, e.g., "Europe/Vienna".}
}
\value{
A character string of the formatted date.
}
\description{
Generates a timestamp string with specific formatting options, focusing on
German natural language formats (e.g., "9. Dezember um 7:20 Uhr").
}
\examples{
dw_timestamp()
dw_timestamp(format = "de_date")
}
