% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_seeded}
\alias{eval_seeded}
\title{reproducible r expressions that include randomness}
\usage{
eval_seeded(expr)
}
\arguments{
\item{expr}{exxpression to evaluate}
}
\description{
reproducible r expressions that include randomness
}
