% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_alias}
\alias{expand_alias}
\title{expand custom specification}
\usage{
expand_alias(spec)
}
\arguments{
\item{spec}{model specification}
}
\value{
\strong{fml}: qualified model specification
}
\description{
expand custom specification
}
