% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{A placeholder function using roxygen}
\usage{
f(x = NULL)
}
\arguments{
\item{x}{An optional character variable, defaults to \sQuote{world}}
}
\value{
return value
}
\description{
This function shows a standard text on the console. In a time-honoured
tradition, it defaults to displaying \emph{hello, world}.
}
