% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{file_entropy}
\alias{file_entropy}
\title{Compute file entropy of an ASCII encoded file}
\usage{
file_entropy(file)
}
\arguments{
\item{file}{path to file}
}
\value{
numeric length 1: the entropy of the file.
}
\description{
Compute file entropy of an ASCII encoded file
}
\examples{
\dontrun{
ascii("testdata/iris.csv")
}
}
