% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaza_advanced.R
\name{gaza_child_name_counts}
\alias{gaza_child_name_counts}
\title{Fetches estimated counts of children killed, grouped by name}
\usage{
gaza_child_name_counts()
}
\value{
A data frame containing the child name count estimates.
}
\description{
Fetches estimated counts of children killed, grouped by name
}
\details{
This dataset is used to derive the estimates for the "killed children by name" visualizations on the Palestine Datasets website. The logic for this calculation can be seen in JavaScript on GitHub. It's available as a JSON API that updates as both the summary dataset and the Killed in Gaza names list receive updates.
}
\examples{
# Get child name counts
child_counts = gaza_child_name_counts()
head(child_counts)

}
