% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaza_killed.R
\name{gaza_killed_v3}
\alias{gaza_killed_v3}
\title{Fetches and processes the v3 list of known individuals killed in Gaza}
\usage{
gaza_killed_v3()
}
\value{
A data frame with daily cumulative deaths, children deaths, and press deaths.
}
\description{
This function retrieves the dataset of names of those killed in Gaza from the v3 API endpoint.
It then processes the data to create a time-series of cumulative deaths based on the release dates.
}
