% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaza_press.R
\name{gaza_press_killed}
\alias{gaza_press_killed}
\title{Fetches the list of journalists killed in Gaza}
\usage{
gaza_press_killed(format = "df", minified = TRUE)
}
\arguments{
\item{format}{A string specifying the return format. Either "df" (default) to get
a data frame from the CSV endpoint or "json" for the raw JSON data.}

\item{minified}{Logical. Whether to return minified JSON. Default is TRUE.}
}
\value{
A data frame or a list containing the data.
}
\description{
Fetches the list of journalists killed in Gaza
}
\details{
\strong{Data Source:}
The file is updated when a new list is released by Gaza's Government Media Office or when incremental updates about new individual incidents are received. These are sourced from the Ministry's Telegram Channel.

The ministry has previously released photos of some of those killed, which can be retrieved in Arabic and English from their public Google Drive. As of writing, it includes photos for just over half of the list.

\strong{Data Fields:}
Each record contains:
\itemize{
\item \code{name}: Original Arabic name from the source list
\item \code{name_en}: English name translation
\item \code{notes}: Includes agency they worked for & available detail on how they were killed
}
}
\examples{
# Get as data frame (recommended)
press_df = gaza_press_killed()

# Get as JSON
press_json = gaza_press_killed(format = "json")

# Get unminified JSON
press_full = gaza_press_killed(format = "json", minified = FALSE)

}
