% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaza_advanced.R
\name{gaza_top_translated_names}
\alias{gaza_top_translated_names}
\title{Fetches the counts of the top 10 translated first names by age group}
\usage{
gaza_top_translated_names()
}
\value{
A list containing the name frequency data, categorized by age group.
}
\description{
Fetches the counts of the top 10 translated first names by age group
}
\details{
This dataset is used to derive estimates of children killed for the home page name cards as documented on the Summary dataset page.
}
\examples{
# Get top translated names by age group
top_names = gaza_top_translated_names()
print(top_names$girl)
print(top_names$boy)
print(top_names$woman)
print(top_names$man)

}
