% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_colors.R
\name{get_best_dst_contrast}
\alias{get_best_dst_contrast}
\title{Get Best Contrasting Color for DST Ressort}
\usage{
get_best_dst_contrast(ressort, contrast_level = "AA", color_type = "all")
}
\arguments{
\item{ressort}{Character. Name of the ressort.}

\item{contrast_level}{Character. Desired contrast level.}

\item{color_type}{Character. Type of contrasting color.}
}
\value{
Character. Hex color code of the best contrasting color.
}
\description{
Quick helper function to get the single best contrasting color for a DST ressort.
}
\examples{
# Get best contrasting color for apo
get_best_dst_contrast("apo")

# Get best dark contrasting color for sport
get_best_dst_contrast("sport", color_type = "dark")

}
