% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{hash}
\alias{hash}
\title{Hash an \R object}
\usage{
hash(asset, algorithm)
}
\arguments{
\item{asset}{an \R object}

\item{algorithm}{xxHash variant to compute.}
}
\value{
\code{chr} an xxHash string of the asset.
}
\description{
\code{dav} uses 64-bit xxh3 (v0.8.1) of Yann Collet's xxHash algorithm to
represent known data asset valuations.
}
\details{
xx
Many hash algorithms exist.

Note that it is NOT cryptographic.
}
\seealso{
\code{\link[=serialize]{base::serialize}} for serializable \R
objects.
}
