% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_bestand.R
\name{kba_bev_anteil}
\alias{kba_bev_anteil}
\title{BEV-Anteil (\%)}
\usage{
kba_bev_anteil(
  geo = c("bundesland", "kreis", "plz", "gemeinde"),
  halter = c("alle", "privat", "gewerblich"),
  file = NULL,
  dir = NULL,
  ...
)
}
\arguments{
\item{geo}{"bundesland", "kreis", "plz", "gemeinde"}

\item{halter}{"alle", "privat", "gewerblich" (nur für PLZ/Gemeinde)}

\item{file}{Pfad zu einer lokalen Excel-Datei (oder \code{NULL} für Auto-Download).}

\item{dir}{Verzeichnis mit mehreren Quartalsdateien (für Zeitreihen).}

\item{...}{Weitere Argumente an \code{\link[=kba_bestand]{kba_bestand()}}.}
}
\value{
Tibble mit zusätzlicher Spalte \code{bev_anteil_pct}
}
\description{
BEV-Anteil (\%)
}
