% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_bestand.R
\name{kba_bev_n}
\alias{kba_bev_n}
\title{Absoluter BEV-Bestand}
\usage{
kba_bev_n(
  geo = c("bundesland", "kreis", "plz", "gemeinde"),
  halter = c("alle", "privat", "gewerblich"),
  file = NULL,
  dir = NULL,
  ...
)
}
\arguments{
\item{geo}{"bundesland", "kreis", "plz", "gemeinde"}

\item{halter}{"alle", "privat", "gewerblich" (nur PLZ/Gemeinde)}

\item{file, dir, ...}{weitergereicht an \code{\link[=kba_bestand]{kba_bestand()}}}
}
\value{
Tibble mit \code{bev_n}
}
\description{
Absoluter BEV-Bestand
}
