% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_bestand.R
\name{kba_download_fz27}
\alias{kba_download_fz27}
\title{Alle KBA FZ27-Excel-Dateien herunterladen}
\usage{
kba_download_fz27(
  dir = "kba_fz27",
  overwrite = FALSE,
  verbose = TRUE,
  page_url = .KBA_FZ27_PAGE_URL
)
}
\arguments{
\item{dir}{Zielverzeichnis.}

\item{overwrite}{Überschreiben?}

\item{verbose}{Fortschrittsmeldungen.}

\item{page_url}{URL der Übersichtsseite.}
}
\value{
Unsichtbarer Vektor mit Pfaden.
}
\description{
Alle KBA FZ27-Excel-Dateien herunterladen
}
