% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_bestand.R
\name{kba_emissionsklassen}
\alias{kba_emissionsklassen}
\title{Emissionsklassen (Euro-Normen) für PKW}
\usage{
kba_emissionsklassen(bundesland = NULL, file = NULL, dir = NULL, ...)
}
\arguments{
\item{bundesland}{Optional einschränken auf ein Bundesland.}

\item{file}{Pfad zu einer lokalen Excel-Datei (oder \code{NULL} für Auto-Download).}

\item{dir}{Verzeichnis mit mehreren Quartalsdateien (für Zeitreihen).}

\item{...}{Weitere Argumente an \code{\link[=kba_bestand]{kba_bestand()}}.}
}
\value{
Tibble mit \code{stichtag}, \code{bundesland}, \code{kraftstoffart}, \code{emissionsklasse}, \code{anzahl}
}
\description{
Emissionsklassen (Euro-Normen) für PKW
}
