% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_bestand.R
\name{kba_non_bev_n}
\alias{kba_non_bev_n}
\title{Absoluter Nicht-BEV-Bestand (alle PKW ohne reine Elektro)}
\usage{
kba_non_bev_n(
  geo = c("bundesland", "kreis", "plz", "gemeinde"),
  halter = c("alle", "privat", "gewerblich"),
  file = NULL,
  dir = NULL,
  ...
)
}
\arguments{
\item{geo}{"bundesland", "kreis", "plz", "gemeinde"}

\item{halter}{"alle", "privat", "gewerblich" (nur PLZ/Gemeinde)}

\item{file, dir, ...}{weitergereicht an \code{\link[=kba_bestand]{kba_bestand()}}}
}
\value{
Tibble mit \code{nicht_bev_n} (= pkw_insgesamt - bev_n, inkl. PHEV/Hybrid/Verbrenner)
}
\description{
Absoluter Nicht-BEV-Bestand (alle PKW ohne reine Elektro)
}
