% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_bestand.R
\name{kba_zeitreihe}
\alias{kba_zeitreihe}
\title{Zeitreihe über mehrere Quartale (nur für Bundesland/Kreis)}
\usage{
kba_zeitreihe(
  geo = c("bundesland", "kreis"),
  antrieb = c("bev", "phev", "hybrid", "gas", "diesel"),
  dir = NULL,
  ...
)
}
\arguments{
\item{geo}{"bundesland" oder "kreis"}

\item{antrieb}{"bev", "phev", "hybrid", "gas", "diesel"}

\item{dir}{Verzeichnis mit mehreren Dateien (erforderlich)}

\item{...}{weitere Argumente an \code{kba_bestand}}
}
\value{
Tibble mit \code{stichtag}, \code{geo_code}, \code{geo_name}, \code{bestand}
}
\description{
Zeitreihe über mehrere Quartale (nur für Bundesland/Kreis)
}
