% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{membership}
\alias{membership}
\title{membership functions}
\description{
membership functions
}
\references{
Talpur, N., Salleh, M. N. M. & Hussain, K. (2017).
An investigation of membership functions on performance of ANFIS for
solving classification problems.
\emph{In IOP Conference Series: Materials Science and Engineering, 226}(1), 012103
}
