% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{mime}
\alias{mime}
\title{package internal mime data}
\description{
A hash table to identify media type of a data asset. See
\href{https://www.iana.org/assignments/media-types/media-types.xhtml}{mime}
}
\details{
A MIME (Multipurpose Internet Mail Extension) describes the nature and format
of a document, file, or assortment of bytes. MIME types are defined and
standardized in IETF's RFC 6838. The Internet Assigned Numbers Authority
(IANA) is responsible for all official MIME types
}
\keyword{dataset}
