% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notnull}
\alias{notnull}
\alias{\%||\%}
\title{null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{a variable}
}
\value{
LHS if true, RHS if LHS is NULL
}
\description{
null coalescing operator
}
