% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_asset}
\alias{sleuth_asset}
\title{Find and organize the file context}
\usage{
sleuth_asset(asset)
}
\arguments{
\item{asset}{data asset}
}
\value{
\strong{lst} of documented file context
}
\description{
Find and organize the file context
}
