% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth_context}
\alias{sleuth_context}
\title{get user and context}
\usage{
sleuth_context()
}
\description{
Several data attributes require detailed information on the who, what, where
and when of data assets. Additionally to facilitate metric reproducibility,
some hard- nd software specifications are \(temporarily\) gathered
}
