% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conds.R
\name{user_interaction}
\alias{user_interaction}
\alias{E}
\alias{W}
\alias{M}
\title{Developer-user interaction tools}
\usage{
E(x, ...)

W(x, ...)

M(x, ...)
}
\arguments{
\item{x}{\tab \code{chr} - condition name}

\item{\dots}{further custom message specifications}
}
\description{
Informative condition handling for known situations and unexpected input
or behavior.
}
\keyword{internal}
