% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_colors.R
\name{visualize_dst_contrasts}
\alias{visualize_dst_contrasts}
\title{Visualize DST Ressort Contrasts}
\usage{
visualize_dst_contrasts(ressort = NULL, contrast_level = "AA", max_options = 6)
}
\arguments{
\item{ressort}{Character. Name of the ressort to visualize. If NULL,
shows all ressorts in a grid.}

\item{contrast_level}{Character. Contrast level to show.}

\item{max_options}{Numeric. Maximum number of contrast options to display per ressort.}
}
\description{
Shows visual comparison of DST ressort colors with their contrasting options.
}
\examples{
# Visualize apo ressort contrasts
visualize_dst_contrasts("apo")

# Show all ressorts
visualize_dst_contrasts()

# Show only high contrast options for sport
visualize_dst_contrasts("sport", contrast_level = "high")

}
