% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{dva}
\alias{dva}
\title{assess the value of data asset}
\usage{
dva(..., formula)
}
\arguments{
\item{...}{the R object to assess the value of}

\item{formula}{a \emph{valuation specification}}
}
\value{
list of length x
}
\description{
assess the value of data asset
}
