% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{is.asset}
\alias{is.asset}
\title{checking if R object or file is a trade-able data asset}
\usage{
is.asset(x, ...)
}
\arguments{
\item{x}{R object or file path}

\item{...}{arguments passed to or from other methods}
}
\value{
boolean
}
\description{
checking if R object or file is a trade-able data asset
}
\details{
a data asset is a collection of relata that carries semantic value in a
context.
}
\note{
\code{is.asset(x)} does not test if the asset is \emph{sensible}.
}
