% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth}
\alias{sleuth}
\title{investigate and document file metadata}
\usage{
\special{sleuth(file)}

\special{sleuth(dir, pattern = ".csv")}
}
\arguments{
\item{asset}{a data asset. A \emph{filepath} or an R object.}

\item{\dots}{arguments passed from and to methods.}
}
\value{
list of file metadata descriptors
}
\description{
\code{sleuth} investigates the context in which data is valued. It lays the
foundation for comparison of data assets and workflows via fuzzy
inference systems.
}
\details{
\code{sleuth} investigates and collects information on the \emph{thing\(s\)}
to be valued and the \emph{system} on which it is evaluated. A comparison of
data.

file paths are case sensitive.
}
\section{Data asset semantics}{

An approximation of data value is limited by the ability of the valuer to
distinguish noise from data. \code{sleuth} documents possible sources
of influence on the integrity of the valuation process. As the package is
open source, sensitive information on the users are not collected.
}

\section{File classification}{

What exactly is a file is operating system dependent. Inference on the
structure of files is based on MIME \(Multipurpose Internet
Mail Extensions\)
}

\examples{
\dontrun{
sleuth(iris)
}
}
