% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth.R
\name{sleuth}
\alias{sleuth}
\title{investigate and document file metadata}
\usage{
\special{dav::sleuth(file)}
}
\arguments{
\item{file}{filepath}
}
\value{
list of file metadata descriptors
}
\description{
\code{dav::sleuth} investigates the context in which data is valued.
}
\details{
\code{sleuth} investigates and collects information on the \emph{thing\(s\)} to be
valued and the \emph{system} on which it is evaluated.
}
